requirements.txt
Flask==2.3.1
Flask-RESTful==0.3.9
pandas==2.0.3
pytest==7.4.0
python-simplexml==0.1.5
requests==2.29.0
requests-toolbelt==0.10.1